refactor: extract route handlers into modules
